Refactor: 파일 구조 변경 및 함수 위치 변경
validateUtil에 들어있던 함수들 중 로그인과 회원가입 페이지에서 공통으로
사용되는 함수는 sign.js파일에 작성.
그 외 함수들은 각자 사용하는 폴더로 이동.
폼이 제출될 때 사용되는 함수는 따로 분리하지 않고 작성.
