+++ signin.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>login</title>
     <link rel="stylesheet" href="./stylesheets/components/sign.css" />
-    <script type="module" src="./scripts/signin.js"></script>
+    <script type="module" src="./scripts/pages/signin.js"></script>
   </head>
   <body>
     <div class="container">
