+++ signup.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>logup</title>
     <link rel="stylesheet" href="./stylesheets/components/sign.css" />
-    <script type="module" src="./scripts/signup.js"></script>
+    <script type="module" src="./scripts/pages/signup.js"></script>
   </head>
   <body>
     <div class="container">
